WIP: refactoring job API
